+++ core/caiyun/options.ts
@@ -8,6 +8,7 @@ export const defuConfig = {
     enable: true,
     digest: '202CB962AC59075B964B07152D234B70',
     inviteCodes: [],
+    waterFriends: [],
   },
   aiRedPack: {
     enable: true,